+++ src/sys/socket/addr.rs
@@ -907,60 +907,39 @@ impl SockAddr {
     /// with the size of the actual data type. sockaddr is commonly used as a proxy for
     /// a superclass as C doesn't support inheritance, so many functions that take
     /// a sockaddr * need to take the size of the underlying type as well and then internally cast it back.
-    pub fn as_ffi_pair(&self) -> (&libc::sockaddr, libc::socklen_t) {
+    pub fn as_ffi_pair(&self) -> (*const libc::sockaddr, libc::socklen_t) {
         match *self {
             SockAddr::Inet(InetAddr::V4(ref addr)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(addr as *const libc::sockaddr_in as *const libc::sockaddr)
-                },
+                addr as *const libc::sockaddr_in as *const libc::sockaddr,
                 mem::size_of_val(addr) as libc::socklen_t
             ),
             SockAddr::Inet(InetAddr::V6(ref addr)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(addr as *const libc::sockaddr_in6 as *const libc::sockaddr)
-                },
+                addr as *const libc::sockaddr_in6 as *const libc::sockaddr,
                 mem::size_of_val(addr) as libc::socklen_t
             ),
             SockAddr::Unix(UnixAddr { ref sun, path_len }) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(sun as *const libc::sockaddr_un as *const libc::sockaddr)
-                },
+                sun as *const libc::sockaddr_un as *const libc::sockaddr,
                 (path_len + offset_of!(libc::sockaddr_un, sun_path)) as libc::socklen_t
             ),
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Netlink(NetlinkAddr(ref sa)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(sa as *const libc::sockaddr_nl as *const libc::sockaddr)
-                },
+                sa as *const libc::sockaddr_nl as *const libc::sockaddr,
                 mem::size_of_val(sa) as libc::socklen_t
             ),
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Alg(AlgAddr(ref sa)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(sa as *const libc::sockaddr_alg as *const libc::sockaddr)
-                },
+                sa as *const libc::sockaddr_alg as *const libc::sockaddr,
                 mem::size_of_val(sa) as libc::socklen_t
             ),
             #[cfg(any(target_os = "ios", target_os = "macos"))]
             SockAddr::SysControl(SysControlAddr(ref sa)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(sa as *const libc::sockaddr_ctl as *const libc::sockaddr)
-                },
+                sa as *const libc::sockaddr_ctl as *const libc::sockaddr,
                 mem::size_of_val(sa) as libc::socklen_t
 
             ),
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Link(LinkAddr(ref addr)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(addr as *const libc::sockaddr_ll as *const libc::sockaddr)
-                },
+                addr as *const libc::sockaddr_ll as *const libc::sockaddr,
                 mem::size_of_val(addr) as libc::socklen_t
             ),
             #[cfg(any(target_os = "dragonfly",
@@ -971,18 +950,12 @@ impl SockAddr {
                       target_os = "netbsd",
                       target_os = "openbsd"))]
             SockAddr::Link(LinkAddr(ref addr)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(addr as *const libc::sockaddr_dl as *const libc::sockaddr)
-                },
+                addr as *const libc::sockaddr_dl as *const libc::sockaddr,
                 mem::size_of_val(addr) as libc::socklen_t
             ),
             #[cfg(any(target_os = "android", target_os = "linux"))]
             SockAddr::Vsock(VsockAddr(ref sa)) => (
-                // This cast is always allowed in C
-                unsafe {
-                    &*(sa as *const libc::sockaddr_vm as *const libc::sockaddr)
-                },
+                sa as *const libc::sockaddr_vm as *const libc::sockaddr,
                 mem::size_of_val(sa) as libc::socklen_t
             ),
         }